Simplify ingestion API key admin: direct DB writes, no proxy
- Add rust/analytics-web-srv/src/ingestion_keys.rs, mirroring
  analytics_keys.rs: mint/list/revoke/import against ingestion_api_keys via
  direct Postgres writes over the shared telemetry-DB pool, gated by the
  AdminUser extractor. Replaces the ingestion_keys_proxy module entirely.
- Remove ingestion's own /auth/api_keys* admin routes
  (rust/public/src/servers/api_keys.rs) and collapse
  serve_ingestion_with_api_key_config back into a single serve_ingestion —
  ingestion now only validates API keys, it exposes no way to
  mint/list/revoke/import them.
- Fix the attribution bug the proxy had: every ingestion-key
  mint/revoke/import now records the acting admin's own OIDC identity
  (created_by/revoked_by), never a shared service credential.
- Drop the second OIDC service credential
  (MICROMEGAS_INGESTION_PROXY_OIDC_*) and MICROMEGAS_INGESTION_ADMIN_URL —
  analytics-web-srv's own admin list is now the single admin surface for
  both key tables.
- Update rust/monolith/src/main.rs for the collapsed serve_ingestion
  signature, dropping the now-unused DbApiKeyConfig construction.
- Change the micromegas-import-keys CLI to always target
  analytics-web-srv's WebClient for both --table values; add
  WebClient.import_ingestion_api_key and remove the direct-to-ingestion
  IngestionClient.
- Add rust/analytics-web-srv/tests/ingestion_keys_tests.rs and a
  make_client unit test on the Python side; update routing_tests.rs and
  test_import_keys.py fixtures for the new module/client shape.
- Update mkdocs docs (api-keys.md, web-app.md, ingestion.md, monolith.md,
  authentication.md, otlp/index.md), docker/README.md,
  docker-compose.monolith.yaml, and the stale frontend/API-key doc
  comments to describe the single direct-write admin surface and the
  narrowed ingestion-role DB grants.
