+++ docker/docker-compose.monolith.yaml
@@ -52,7 +52,7 @@ services:
       #   MICROMEGAS_INGESTION_API_KEYS  or  MICROMEGAS_API_KEYS  (transitional env keyring),
       #   or a non-empty ingestion_api_keys / analytics_api_keys DB table (see
       #   mkdocs/docs/admin/api-keys.md — the steady-state path, minted via
-      #   POST /auth/api_keys once the ingestion role is up)
+      #   the web role's POST /api/ingestion-api-keys / POST /api/analytics-api-keys)
       #   MICROMEGAS_ANALYTICS_OIDC_CONFIG  or  MICROMEGAS_OIDC_CONFIG
     volumes:
       - lake:/data
