+++ mkdocs/docs/admin/authentication.md
@@ -109,11 +109,11 @@ The `MICROMEGAS_ADMINS` environment variable is a JSON array of user identifiers
 
 ### API Key Configuration
 
-**Steady state: DB-backed keys.** Mint an ingestion key with `POST /auth/api_keys`
-on the ingestion service, or an analytics key with `POST /api/analytics-api-keys`
-on `analytics-web-srv` (both OIDC + admin required) — see
-[API Keys](api-keys.md) for the full route reference and the `mmk_`-prefixed
-key shape.
+**Steady state: DB-backed keys.** Mint an ingestion key with
+`POST /api/ingestion-api-keys`, or an analytics key with
+`POST /api/analytics-api-keys` — both on `analytics-web-srv` (OIDC + admin
+required) — see [API Keys](api-keys.md) for the full route reference and the
+`mmk_`-prefixed key shape.
 
 **Legacy/bootstrap: the env keyring.** Still the only option for
 `object-cache-srv` (env-only permanently — see [Object Cache](object-cache.md)),
@@ -672,8 +672,8 @@ the `object-cache-srv` exception, which never migrates).
 2. **Populate the tables** from the existing env keyring using the
    `micromegas-import-keys` CLI tool (see
    [Migrating from the env keyring](api-keys.md#migrating-from-the-env-keyring)
-   for the exact commands), or mint fresh keys via `POST /auth/api_keys` for
-   callers you can update. A key valid on both ingestion and flight-sql today
+   for the exact commands), or mint fresh keys via `POST /api/ingestion-api-keys`
+   on `analytics-web-srv` for callers you can update. A key valid on both ingestion and flight-sql today
    must become two distinct keys — "never both" is enforced at the code level
    once the tables are in use.
 3. **Set up an OIDC provider** (Google, Azure AD, etc.) for human/service-account
