+++ rust/analytics-web-srv/src/lib.rs
@@ -6,7 +6,7 @@ pub mod auth;
 pub mod data_source_cache;
 pub mod data_sources;
 pub mod folders;
-pub mod ingestion_keys_proxy;
+pub mod ingestion_keys;
 pub mod maps;
 pub mod screen_types;
 pub mod screens;
