+++ rust/monolith/src/main.rs
@@ -23,13 +23,13 @@ use anyhow::{Context, Result};
 use clap::Parser;
 use micromegas::analytics::lakehouse::lakehouse_context::LakehouseContext;
 use micromegas::analytics::lakehouse::view_factory::default_view_factory;
-use micromegas::auth::db_api_key::{ApiKeyTable, DbApiKeyConfig, dedicated_key_store_pool};
+use micromegas::auth::db_api_key::{ApiKeyTable, dedicated_key_store_pool};
 use micromegas::auth::default_provider::ProviderBuilder;
 use micromegas::ingestion::data_lake_config::DataLakeConfig;
 use micromegas::ingestion::remote_data_lake::connect_to_remote_data_lake;
 use micromegas::micromegas_main;
 use micromegas::servers::flight_sql_server::FlightSqlServer;
-use micromegas::servers::ingestion::serve_ingestion_with_api_key_config;
+use micromegas::servers::ingestion::serve_ingestion;
 use micromegas::servers::maintenance::{daemon, get_global_views_with_update_group};
 use micromegas::servers::shutdown::{ShutdownFanout, wait_for_sigterm};
 use micromegas::tracing::prelude::*;
@@ -267,18 +267,9 @@ async fn main() -> Result<()> {
         let listen_addr = args.listen_endpoint_http;
         let grace_c = grace;
         let auth = ingestion_auth;
-        let api_key_config = DbApiKeyConfig::from_env_with_prefix("MICROMEGAS_INGESTION");
-        join_set.spawn(async move {
-            serve_ingestion_with_api_key_config(
-                listen_addr,
-                lake,
-                auth,
-                shutdown,
-                grace_c,
-                api_key_config,
-            )
-            .await
-        });
+        join_set.spawn(
+            async move { serve_ingestion(listen_addr, lake, auth, shutdown, grace_c).await },
+        );
     }
 
     // ── FlightSQL ──────────────────────────────────────────────────────────
