+++ rust/public/src/servers/mod.rs
@@ -53,9 +53,6 @@ pub mod perfetto;
 /// metadata about this implementation of FlightSQL
 pub mod sqlinfo;
 
-/// DB-backed API key management HTTP routes (mint/list/revoke, #1383)
-pub mod api_keys;
-
 /// log uris of http requests
 pub mod log_uri_service;
